refactor: shrink modules under size budget via extraction + import collapse
