+++ src/synthorg/observability/prometheus_labels.py
@@ -18,62 +18,6 @@ from synthorg.observability import get_logger
 from synthorg.observability.events.metrics import METRICS_SCRAPE_FAILED
 from synthorg.providers.errors import ProviderErrorLabel
 
-__all__ = [
-    "MCP_UNKNOWN_TOOL_LABEL",
-    "TRANSIENT_PROVIDER_ERROR_CLASSES",
-    "UNKNOWN_LABEL",
-    "VALID_API_ERROR_CATEGORIES",
-    "VALID_APPROVAL_OUTCOMES",
-    "VALID_AUDIT_APPEND_STATUSES",
-    "VALID_AUDIT_VERIFICATION_OUTCOMES",
-    "VALID_AUTONOMY_PROMOTION_OUTCOMES",
-    "VALID_BLUEPRINT_OUTCOMES",
-    "VALID_BUDGET_QUERY_TYPES",
-    "VALID_CACHE_NAMES",
-    "VALID_CACHE_OUTCOMES",
-    "VALID_DISCONNECT_REASONS",
-    "VALID_DISCONNECT_TRANSPORTS",
-    "VALID_ESCALATION_OUTCOMES",
-    "VALID_IDENTITY_CHANGE_TYPES",
-    "VALID_LOG_SINK_KINDS",
-    "VALID_LOG_SINK_OUTCOMES",
-    "VALID_MCP_HANDLER_OUTCOMES",
-    "VALID_OTLP_KINDS",
-    "VALID_OTLP_OUTCOMES",
-    "VALID_PG_BACKENDS",
-    "VALID_PROVIDER_CALL_TYPES",
-    "VALID_PROVIDER_ERROR_CLASSES",
-    "VALID_PUSH_QUEUE_OUTCOMES",
-    "VALID_SETTINGS_NAMESPACES",
-    "VALID_STATUS_CLASSES",
-    "VALID_TASK_OUTCOMES",
-    "VALID_TOKEN_DIRECTIONS",
-    "VALID_TOOL_OUTCOMES",
-    "VALID_VERDICTS",
-    "VALID_WORKFLOW_EXECUTION_STATUSES",
-    "VALID_WS_REVALIDATION_OUTCOMES",
-    "VALID_WS_TRANSPORTS",
-    "_LabelSnapshot",
-    "_reset_label_snapshot_for_tests",
-    "_reset_mcp_tool_names_for_tests",
-    "_snapshot_lock",
-    "is_known_agent_id",
-    "normalize_mcp_tool_label",
-    "normalize_model_label",
-    "normalize_provider_label",
-    "register_mcp_tool_names",
-    "require_finite",
-    "require_label",
-    "require_label_summary",
-    "require_non_negative",
-    "status_class",
-    "update_label_snapshot",
-    "validate_agent_id",
-    "validate_department",
-    "validate_tool_name",
-    "validate_workflow_definition_id",
-]
-
 logger = get_logger(__name__)
